requirements.txt
requests
feedparser
python-dotenv
pandas
schedule
schedule
